+++ src/lib/codegen-graphql.ts
@@ -100,15 +100,22 @@ function generateGraphQlCodeSpread(ty: SpreadAssertion, ctx: CodegenContext) {
 
 
 function generateGraphQlCodeSequence(ty: SequenceAssertion, ctx: CodegenContext) {
-    return 'Any'; // TODO: Any is invalid type.
+    return '[Any!]'; // TODO: Any is invalid type.
 }
 
 
 function generateGraphQlCodeOneOf(ty: OneOfAssertion, ctx: CodegenContext) {
-    return `${ty.oneOf
-        .map(x => x.typeName ?
-            x.typeName :
-            generateGraphQlCodeInner(x, false, ctx)).join(' | ')}`;
+    const filtered = ty.oneOf.filter(x => !(
+        x.kind === 'primitive' && (x.primitiveName === 'null' || x.primitiveName === 'undefined') ||
+        x.kind === 'primitive-value' && (x.value === null || x.value === void 0)));
+    if (filtered.length === 1 && ty.oneOf.length !== 1) {
+        return generateGraphQlCodeInner(filtered[0], false, ctx);
+    } else {
+        return `${ty.oneOf
+            .map(x => x.typeName ?
+                x.typeName :
+                generateGraphQlCodeInner(x, false, ctx)).join(' | ')}`;
+    }
 }